requirements.txt
streamlit
qrcode
Pillow
openpyxl
pandas 